fix(regex): keep the UTF-16 index helpers outside the regex-engine gate
The UTF-16 <-> byte conversion helpers lived in `exec_array`, which is behind
`#[cfg(feature = "regex-engine")]` — but `replace_fn` and `match_string` are
NOT gated and now call them, so a `--no-default-features` build failed with

  error[E0433]: cannot find `exec_array` in `super`

which is what broke compiler-output-regression's "Build compiler" step.

The helpers are pure UTF-16 index math with no regex dependency, so they move
to an ungated `regex::utf16` module; `exec_array` re-exports them for the
gated call sites.

Same cross-gate shape as #6303: a `#[cfg]`-gated callee reached from an
ungated caller. Verified all three configs build clean: regex-engine off,
regex-engine on, and the full `-p perry` compiler.
